change default gear selection from PARK on NEUTRAL
The default selection is transmitted when the real state is not ready.
For example, if GPIO driver is not loaded the default state will be
transmitted. It may lead to undesired side-effects, for example: if
PARK mode is transmitted then EVS application will be enabled.

It is proposed to change the default state to eliminate any side-effects.

Test: disable GPIO driver in kernel configs `CONFIG_KEYBOARD_GPIO=n`,
      build image and flush a DUT, boot and ensure that there are not any
      undesired side-effects.
Jira-Id: ADM-3914
Change-Id: I3e6188d2d156dd12e0d050d67dd00e79e30faeb1
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>